feat: adjusted admin node management
