+++ packages/client-ui/components/layout/header.vue
@@ -30,7 +30,7 @@ export default defineNuxtComponent({
             displayNav.value = !displayNav.value;
         };
 
-        const infoText = computed(() => realmManagementName.value || 'Personal Health Train');
+        const infoText = computed(() => realmManagementName.value || 'PrivateAim');
 
         return {
             infoText,
@@ -59,7 +59,7 @@ export default defineNuxtComponent({
                     </button>
                 </div>
                 <div class="logo">
-                    <span>U</span>I
+                    <span>P</span>o<span>r</span>t<span>a</span>l
                     <span class="info-text">{{ infoText }}</span>
                 </div>
             </div>
