+++ packages/client-ui/package.json
@@ -11,8 +11,7 @@
     "description": "This package contains the frontend application of the HUB Ecosystem.",
     "scripts": {
         "dev": "nuxi dev",
-        "build": "echo 1",
-        "build-alt": "nuxi build",
+        "build": "nuxi build",
         "start": "nuxi start",
         "generate": "nuxi generate",
         "prepare": "nuxi prepare",
